+++ libraries/ui-library/src/components/six-file-upload/test/six-file-upload.spec.tsx
@@ -15,7 +15,7 @@ describe('six-file-upload', () => {
     expect(page.root).toEqualHtml(`
         <six-file-upload>
           <div class="six-file-upload">
-            <six-card class="six-file-upload__container--full">
+            <six-card aria-invalid="false" class="six-file-upload__container--full">
               <div class="six-file-upload__drop-zone">
                 <div>
                   <span>
@@ -30,6 +30,7 @@ describe('six-file-upload', () => {
                 </div>
               </div>
             </six-card>
+            <div aria-hidden="true"></div>
           </div>
         </six-file-upload>
     `);
@@ -44,7 +45,7 @@ describe('six-file-upload', () => {
     expect(page.root).toEqualHtml(`
         <six-file-upload label="some custom label">
           <div class="six-file-upload">
-            <six-card class="six-file-upload__container--full">
+            <six-card aria-invalid="false" class="six-file-upload__container--full">
               <div class="six-file-upload__drop-zone">
                 <div>
                   <span>
@@ -54,6 +55,7 @@ describe('six-file-upload', () => {
                 </div>
               </div>
             </six-card>
+            <div aria-hidden="true"></div>
           </div>
         </six-file-upload>
     `);
@@ -68,7 +70,7 @@ describe('six-file-upload', () => {
     expect(page.root).toEqualHtml(`
         <six-file-upload disabled="">
           <div class="six-file-upload six-file-upload--disabled">
-            <six-card class="six-file-upload__container--full" disabled="">
+            <six-card aria-invalid="false" class="six-file-upload__container--full" disabled="">
               <div class="six-file-upload__drop-zone">
                 <div>
                   <span>
@@ -83,6 +85,7 @@ describe('six-file-upload', () => {
                 </div>
               </div>
             </six-card>
+            <div aria-hidden="true"></div>
           </div>
         </six-file-upload>
     `);
@@ -97,7 +100,7 @@ describe('six-file-upload', () => {
     expect(page.root).toEqualHtml(`
         <six-file-upload compact="">
           <div class="six-file-upload">
-            <six-button class="six-file-upload__container--compact">
+            <six-button aria-invalid="false" class="six-file-upload__container--compact">
               <span slot="prefix">
                 <six-icon class="six-file-upload__label-icon">
                   arrow_circle_up
@@ -112,6 +115,7 @@ describe('six-file-upload', () => {
                 </div>
               </div>
             </six-button>
+            <div aria-hidden="true"></div>
           </div>
         </six-file-upload>
     `);
@@ -126,7 +130,7 @@ describe('six-file-upload', () => {
     expect(page.root).toEqualHtml(`
         <six-file-upload compact="" disabled="">
           <div class="six-file-upload six-file-upload--disabled">
-            <six-button class="six-file-upload__container--compact" disabled="">
+            <six-button aria-invalid="false" class="six-file-upload__container--compact" disabled="">
               <span slot="prefix">
                 <six-icon class="six-file-upload__label-icon">
                   arrow_circle_up
@@ -141,6 +145,7 @@ describe('six-file-upload', () => {
                 </div>
               </div>
             </six-button>
+            <div aria-hidden="true"></div>
           </div>
         </six-file-upload>
     `);
@@ -155,7 +160,7 @@ describe('six-file-upload', () => {
     expect(page.root).toEqualHtml(`
         <six-file-upload uploading="">
           <div class="six-file-upload six-file-upload--disabled">
-            <six-card class="six-file-upload__container--full" disabled="">
+            <six-card aria-invalid="false" class="six-file-upload__container--full" disabled="">
               <div class="six-file-upload__drop-zone">
                 <span class="six-file-upload__drop-zone__spinner-container">
                   <six-spinner></six-spinner>
@@ -163,6 +168,115 @@ describe('six-file-upload', () => {
                 </span>
               </div>
             </six-card>
+            <div aria-hidden="true"></div>
+          </div>
+          </six-file-upload>
+    `);
+  });
+
+  it('renders invalid', async () => {
+    const page = await newSpecPage({
+      components: [SixFileUpload],
+      html: `<six-file-upload invalid error-text="error message"></six-file-upload>`,
+    });
+
+    expect(page.root).toEqualHtml(`
+        <six-file-upload invalid error-text="error message">
+          <div class="six-file-upload">
+            <six-card aria-invalid="true" class="six-file-upload__container--full">
+              <div class="six-file-upload__drop-zone">
+                <div>
+                  <span>
+                    <span>
+                      Drop files to upload, or
+                      <span class="six-file-upload__label--highlighted">
+                        browse
+                      </span>
+                    </span>
+                  </span>
+                  <input class="six-file-upload__input" name="resume" type="file">
+                </div>
+              </div>
+            </six-card>
+            <div aria-hidden="false">
+            <slot-fb name="error-text">
+              <six-error>
+                <div class="six-file-upload__error-text">error message</div>
+              </six-error>
+            </slot-fb>
+            </div>
+          </div>
+        </six-file-upload>
+    `);
+  });
+
+  it('does not render error message if not invalid', async () => {
+    const page = await newSpecPage({
+      components: [SixFileUpload],
+      html: `<six-file-upload error-text="error message"></six-file-upload>`,
+    });
+
+    expect(page.root).toEqualHtml(`
+        <six-file-upload error-text="error message">
+          <div class="six-file-upload">
+            <six-card aria-invalid="false" class="six-file-upload__container--full">
+              <div class="six-file-upload__drop-zone">
+                <div>
+                  <span>
+                    <span>
+                      Drop files to upload, or
+                      <span class="six-file-upload__label--highlighted">
+                        browse
+                      </span>
+                    </span>
+                  </span>
+                  <input class="six-file-upload__input" name="resume" type="file">
+                </div>
+              </div>
+            </six-card>
+            <div aria-hidden="true">
+              <slot-fb name="error-text">
+                <six-error>
+                  <div class="six-file-upload__error-text">error message</div>
+                </six-error>
+              </slot-fb>
+            </div>
+          </div>
+        </six-file-upload>
+    `);
+  });
+
+  it('render error slot when invalid', async () => {
+    const page = await newSpecPage({
+      components: [SixFileUpload],
+      html: `<six-file-upload invalid><slot name="error-text"><b>bold error</b></six-file-upload>`,
+    });
+
+    expect(page.root).toEqualHtml(`
+        <six-file-upload invalid>
+          <slot hidden="" name="error-text">
+            <b>
+              bold error
+            </b>
+          </slot>
+          <div class="six-file-upload">
+            <six-card aria-invalid="true" class="six-file-upload__container--full">
+              <div class="six-file-upload__drop-zone">
+                <div>
+                  <span>
+                    <span>
+                      Drop files to upload, or
+                      <span class="six-file-upload__label--highlighted">
+                        browse
+                      </span>
+                    </span>
+                  </span>
+                  <input class="six-file-upload__input" name="resume" type="file">
+                </div>
+              </div>
+            </six-card>
+            <div aria-hidden="false">
+            </div>
           </div>
         </six-file-upload>
     `);